+++ cookbooks/arcgis-enterprise/resources/portal.rb
@@ -22,9 +22,9 @@ unified_mode true
 actions :system, :unpack, :install, :uninstall, :stop, :start,
         :update_account, :configure_autostart, :authorize,
         :create_site, :join_site, :set_system_properties, :configure_https,
-        :unregister_standby, :register_server, :federate_server,
-        :unfederate_server, :enable_server_function, :set_allssl,
-        :set_identity_store, :configure_hostidentifiers_properties,
+        :unregister_machine, :unregister_standby, :register_server,
+        :federate_server, :unfederate_server, :enable_server_function,
+        :set_allssl, :set_identity_store, :configure_hostidentifiers_properties,
         :import_root_cert, :webgisdr_export, :webgisdr_import
 
 attribute :setup_archive, :kind_of => String