+++ cookbooks/arcgis-notebooks/resources/server.rb
@@ -33,6 +33,7 @@ attribute :run_as_password, :kind_of => String, :sensitive => true
 attribute :run_as_msa, :kind_of => [TrueClass, FalseClass], :default => false
 attribute :authorization_file, :kind_of => String
 attribute :authorization_file_version, :kind_of => String
+attribute :authorization_options, :kind_of => String, :default => ''
 attribute :product_code, :kind_of => String
 attribute :username, :kind_of => String
 attribute :password, :kind_of => String, :sensitive => true
@@ -259,9 +260,9 @@ action :configure_autostart do
       source agsnotebook_service_file
       cookbook 'arcgis-notebooks'
       variables agsnotebook_template_variables
-      owner agsuser
+      owner 'root'
       group 'root'
-      mode '0755'
+      mode '0600'
       notifies :run, 'execute[Load agsnotebook systemd unit file]', :immediately
     end
 
@@ -288,7 +289,7 @@ action :authorize do
     cmd = node['arcgis']['notebook_server']['authorization_tool']
 
     if node['platform'] == 'windows'
-      args = "/VER #{@new_resource.authorization_file_version} /LIF \"#{@new_resource.authorization_file}\" /S"
+      args = "/VER #{@new_resource.authorization_file_version} /LIF \"#{@new_resource.authorization_file}\" /S #{@new_resource.authorization_options}"
       sa_cmd = Mixlib::ShellOut.new("\"#{cmd}\" #{args}", { :timeout => 600 })
 
       sleep(rand(0..120)) # Use random delay top reduce probability of multiple machines authorization at the same time
@@ -313,7 +314,7 @@ action :authorize do
   
       sa_cmd.error!
     else
-      args = "-f \"#{@new_resource.authorization_file}\""
+      args = "-f \"#{@new_resource.authorization_file}\" #{@new_resource.authorization_options}"
       sa_cmd = Mixlib::ShellOut.new("\"#{cmd}\" #{args}",
             { :timeout => 600, :user => node['arcgis']['run_as_user'] })
       sleep(rand(0..120)) # Use random delay top reduce probability of multiple machines authorization at the same time