+++ cookbooks/esri-tomcat/metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Installs/Configures esri-tomcat'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.3.0'
+version          '0.3.1'
 chef_version     '>= 15.3'
 
 depends          'tomcat', '~> 5.0'