+++ cookbooks/arcgis-insights/CHANGELOG.md
@@ -2,47 +2,69 @@
 
 This file is used to list changes made in each version of the arcgis-insights cookbook.
 
+## 5.1.0
+
+- Added support for ArcGIS Insights 2024.1.
+
+## 5.0.0
+
+- Added support for ArcGIS Insights 2023.2/2023.3.
+
 ## 4.2.0
+
 - Added support for ArcGIS Insights 2023.1.
 
 ## 4.1.0
+
 - Added support for ArcGIS Insights 2022.3.
 
 ## 4.0.0
+
 - Added support for ArcGIS Insights 2022.1/2022.1.1/2022.2.
 
 ## 3.8.0
+
 - Added support for Insights for ArcGIS 2021.1/2021.1.1/2021.2/2021.2.1/2021.3/2021.3.1.
 
 ## 3.7.0
+
 - Added support for Insights for ArcGIS 2020.3.
 
 ## 3.6.1
+
 - Added support for Chef Client 15.
 - Added support for Insights for ArcGIS 2020.2.
 
 ## 3.6.0
+
 - Added support for Insights for ArcGIS 3.4/3.4.1/2020.1
 
 ## 3.5.0
+
 - Added support for Insights for ArcGIS 3.3/3.3.1
 
 ## 3.4.0
+
 - Added support for Insights for ArcGIS 3.2/3.2.1
 
 ## 3.3.0
+
 - Added support for Insights for ArcGIS 3.0/3.1
 - Added dependency on new arcgis-repository cookbook to support global ArcGIS software repository in S3.
 
 ## 3.2.1
+
 - Added support for Insights for ArcGIS 2.2
 - Add ec2 test kitchen.
 
 ## 3.2.0
+
 - Added support for Insights for ArcGIS 2.0/2.1
 
 ## 3.1.0
+
 - Added support for Insights for ArcGIS 1.2
 
 ## 3.0.0
+
 - Initial release of arcgis-insights cookbook