+++ cookbooks/arcgis-mission/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of the arcgis-mission cookbook.
 
+# 5.1.0
+
+- Added support for ArcGIS Mission Server 11.4.
+
 # 5.0.0
 
 - Added support for ArcGIS Mission Server 11.3.